requirements.txt
torch
mujoco
scipy
numpy==1.24
loguru
pyyaml
matplotlib
opencv-python
rerun